dailies: A package for working with the RStudio daily build pages.

RStudio offer their daily builds of the IDE to the community so that we can help out with bug reports and check out new features before they hit the stable releases. The RStudio dailies website is available at https://dailies.rstudio.com

Details

The dailies package provides several functions for working with the dailies page itself and is intended to make it easier for people to get useful URLs from within R.

Important Note

Some functions request data from the RStudio website. Please be respectful in your usage of these functions. The dailies site is offered as a service to the community and we should not repay that kindness by hammering the site.
This package is very much "unofficial".

dailies functions

The most useful function is latest() which returns the url for the latest version.